+++ question/tests/events_test.php
@@ -40,166 +40,6 @@ class core_question_events_testcase extends advanced_testcase {
     }
 
     /**
-     * Test the question category created event.
-     */
-    public function test_question_category_created() {
-        $this->setAdminUser();
-        $course = $this->getDataGenerator()->create_course();
-        $quiz = $this->getDataGenerator()->create_module('quiz', array('course' => $course->id));
-
-        $contexts = new question_edit_contexts(context_module::instance($quiz->cmid));
-
-        $defaultcategoryobj = question_make_default_categories(array($contexts->lowest()));
-        $defaultcategory = $defaultcategoryobj->id . ',' . $defaultcategoryobj->contextid;
-
-        $qcobject = new question_category_object(
-            1,
-            new moodle_url('/mod/quiz/edit.php', array('cmid' => $quiz->cmid)),
-            $contexts->having_one_edit_tab_cap('categories'),
-            $defaultcategoryobj->id,
-            $defaultcategory,
-            null,
-            $contexts->having_cap('moodle/question:add'));
-
-        // Trigger and capture the event.
-        $sink = $this->redirectEvents();
-        $categoryid = $qcobject->add_category($defaultcategory, 'newcategory', '', true);
-        $events = $sink->get_events();
-        $event = reset($events);
-
-        // Check that the event data is valid.
-        $this->assertInstanceOf('\core\event\question_category_created', $event);
-        $this->assertEquals(context_module::instance($quiz->cmid), $event->get_context());
-        $expected = array($course->id, 'quiz', 'addcategory', 'view.php?id=' . $quiz->cmid , $categoryid, $quiz->cmid);
-        $this->assertEventLegacyLogData($expected, $event);
-        $this->assertEventContextNotUsed($event);
-    }
-
-    /**
-     * Test the question category deleted event.
-     */
-    public function test_question_category_deleted() {
-        $this->setAdminUser();
-        $course = $this->getDataGenerator()->create_course();
-        $quiz = $this->getDataGenerator()->create_module('quiz', ['course' => $course->id]);
-
-        $contexts = new question_edit_contexts(context_module::instance($quiz->cmid));
-
-        $defaultcategoryobj = question_make_default_categories([$contexts->lowest()]);
-        $defaultcategory = $defaultcategoryobj->id . ',' . $defaultcategoryobj->contextid;
-
-        $qcobject = new question_category_object(
-                1,
-                new moodle_url('/mod/quiz/edit.php', ['cmid' => $quiz->cmid]),
-                $contexts->having_one_edit_tab_cap('categories'),
-                $defaultcategoryobj->id,
-                $defaultcategory,
-                null,
-                $contexts->having_cap('moodle/question:add'));
-
-        // Create the category.
-        $categoryid = $qcobject->add_category($defaultcategory, 'newcategory', '', true);
-
-        // Trigger and capture the event.
-        $sink = $this->redirectEvents();
-        $qcobject->delete_category($categoryid);
-        $events = $sink->get_events();
-        $event = reset($events);
-
-        // Check that the event data is valid.
-        $this->assertInstanceOf('\core\event\question_category_deleted', $event);
-        $this->assertEquals(context_module::instance($quiz->cmid), $event->get_context());
-        $this->assertEquals($categoryid, $event->objectid);
-        $this->assertDebuggingNotCalled();
-    }
-
-    /**
-     * Test the question category updated event.
-     */
-    public function test_question_category_updated() {
-        $this->setAdminUser();
-        $course = $this->getDataGenerator()->create_course();
-        $quiz = $this->getDataGenerator()->create_module('quiz', ['course' => $course->id]);
-
-        $contexts = new question_edit_contexts(context_module::instance($quiz->cmid));
-
-        $defaultcategoryobj = question_make_default_categories([$contexts->lowest()]);
-        $defaultcategory = $defaultcategoryobj->id . ',' . $defaultcategoryobj->contextid;
-
-        $qcobject = new question_category_object(
-                1,
-                new moodle_url('/mod/quiz/edit.php', ['cmid' => $quiz->cmid]),
-                $contexts->having_one_edit_tab_cap('categories'),
-                $defaultcategoryobj->id,
-                $defaultcategory,
-                null,
-                $contexts->having_cap('moodle/question:add'));
-
-        // Create the category.
-        $categoryid = $qcobject->add_category($defaultcategory, 'newcategory', '', true);
-
-        // Trigger and capture the event.
-        $sink = $this->redirectEvents();
-        $qcobject->update_category($categoryid, $defaultcategory, 'updatedcategory', '', FORMAT_HTML, '', false);
-        $events = $sink->get_events();
-        $event = reset($events);
-
-        // Check that the event data is valid.
-        $this->assertInstanceOf('\core\event\question_category_updated', $event);
-        $this->assertEquals(context_module::instance($quiz->cmid), $event->get_context());
-        $this->assertEquals($categoryid, $event->objectid);
-        $this->assertDebuggingNotCalled();
-    }
-
-    /**
-     * Test the question category viewed event.
-     * There is no external API for viewing the category, so the unit test will simply
-     * create and trigger the event and ensure data is returned as expected.
-     */
-    public function test_question_category_viewed() {
-
-        $this->setAdminUser();
-        $course = $this->getDataGenerator()->create_course();
-        $quiz = $this->getDataGenerator()->create_module('quiz', ['course' => $course->id]);
-
-        $contexts = new question_edit_contexts(context_module::instance($quiz->cmid));
-
-        $defaultcategoryobj = question_make_default_categories([$contexts->lowest()]);
-        $defaultcategory = $defaultcategoryobj->id . ',' . $defaultcategoryobj->contextid;
-
-        $qcobject = new question_category_object(
-                1,
-                new moodle_url('/mod/quiz/edit.php', ['cmid' => $quiz->cmid]),
-                $contexts->having_one_edit_tab_cap('categories'),
-                $defaultcategoryobj->id,
-                $defaultcategory,
-                null,
-                $contexts->having_cap('moodle/question:add'));
-
-        // Create the category.
-        $categoryid = $qcobject->add_category($defaultcategory, 'newcategory', '', true);
-
-        // Log the view of this category.
-        $category = new stdClass();
-        $category->id = $categoryid;
-        $context = context_module::instance($quiz->cmid);
-        $event = \core\event\question_category_viewed::create_from_question_category_instance($category, $context);
-
-        // Trigger and capture the event.
-        $sink = $this->redirectEvents();
-        $event->trigger();
-        $events = $sink->get_events();
-        $event = reset($events);
-
-        // Check that the event data is valid.
-        $this->assertInstanceOf('\core\event\question_category_viewed', $event);
-        $this->assertEquals(context_module::instance($quiz->cmid), $event->get_context());
-        $this->assertEquals($categoryid, $event->objectid);
-        $this->assertDebuggingNotCalled();
-
-    }
-
-    /**
      * Test the questions imported event.
      * There is no easy way to trigger this event using the API, so the unit test will simply
      * create and trigger the event and ensure data is returned as expected.
